20200902 21:22 AWS Web Services

+++ src/main/java/com/muffin/web/asset/AssetServiceImpl.java
@@ -324,8 +324,8 @@ public class AssetServiceImpl implements AssetService {
     }
 
     @Override
-    public int count() {
-        return (int) repository.count();
+    public Long count() {
+        return repository.count();
     }
 
     @Override
