+++ src/main/java/com/muffin/web/board/BoardServiceImpl.java
@@ -108,8 +108,8 @@ public class BoardServiceImpl implements BoardService {
     }
 
     @Override
-    public int count() {
-        return (int)repository.count();
+    public Long count() {
+        return repository.count();
     }
 
     @Override
